Allow for no liquidity limit order placement (#2854)
# Description

We currently require all orders to have quote in order to accurately
compute fee policies (price improvement policies requires quote data).
In the absence of the quote we don't allow order placement and exclude
the order from solvable orders (inconsistently on a one-off code path
outside of the solvable orders cache where it doesn't get flagged as
"filtered" or "invalid" thus being silently ignored).

This seems unnecessarily restrictive. In the absence of a quote we could
simply assume that the limit price itself was the quote (ie apply a pure
surplus fee policy) to the order. This would allow placing limit orders
for tokens which are not yet tradables but will become tradable soon.

@sunce86 am I missing any other reason why we need a quote?

# Changes

- [x] Allow order placement if a quote wasn't found due to a pricing
error
- [x] Don't filter out orders from the auction for which we don't have a
quote
- [x] Assume the limit amounts are equal to the quote amounts for fee
policy application in case we cannot find a quote

## How to test
Introduced an e2e test showing that this works